requirements.txt
youtube-transcript-api>=0.6.1
google-api-python-client>=2.0.0
python-dotenv>=1.0.0